openmp for loop counters must be signed

+++ include/opengm/learning/dataset/dataset.hxx
@@ -152,8 +152,8 @@ namespace opengm {
     typename GM::ValueType Dataset<GM, LOSS, LOSS_GM>::getTotalLossParallel(const typename INF::Parameter& para) const {
         double totalLoss = 0;
         #pragma omp parallel for reduction(+:totalLoss)  
-        for(size_t i=0; i<this->getNumberOfModels(); ++i) {
-            totalLoss = totalLoss + this->getLoss<INF>(para, i);
+        for(long long i=0; i<this->getNumberOfModels(); ++i) {
+            totalLoss = totalLoss + this->getLoss<INF>(para, (size_t)i);
         }
         return totalLoss;
     }
